feat: major refactor complete
