+++ components/HeaderInfo.tsx
@@ -1,0 +1,7 @@
+const HeaderInfo = ({ label, value }: { label: string; value: string }) => (
+  <p>
+    <span className="font-bold">{label}:</span> {value}
+  </p>
+)
+
+export default HeaderInfo
